Update Basic setup functions and read functions for ds18b20
Low priority: threaded reads. This would perhaps be better implemented in
a library where this driver(?) will be used in.
Note: due to how incredably slow the ds18b20 is to read, its highly reccomented
to manage their reads on a seperate thread. (rather, a thread only looking at
the w1 interface)